+++ apps/dashboard/src/components/billing/utils/action.button.constants.ts
@@ -1,0 +1,4 @@
+export enum ActionType {
+  BUTTON = 'button',
+  CONTACT = 'contact',
+}
